更新：优化 LOC，并使用 KnownColor 替换硬编码颜色
