+++ DailyRoutines.ModulesPublic.sln
@@ -1,0 +1,24 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.2.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DailyRoutines.ModulesPublic", "DailyRoutines.ModulesPublic.csproj", "{3FDE884F-5963-1C1A-7C2C-C3B7B7B385F0}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3FDE884F-5963-1C1A-7C2C-C3B7B7B385F0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3FDE884F-5963-1C1A-7C2C-C3B7B7B385F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3FDE884F-5963-1C1A-7C2C-C3B7B7B385F0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3FDE884F-5963-1C1A-7C2C-C3B7B7B385F0}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {4942CBEE-B83C-4CAE-B9DD-250A5D85C4D9}
+	EndGlobalSection
+EndGlobal
